requirements.txt
aioconsole==0.3.3
cycler==0.11.0
fonttools==4.28.5
kiwisolver==1.3.2
llvmlite==0.37.0
matplotlib==3.5.1
numba==0.54.1
numpy==1.20.3
packaging==21.3
Pillow==9.0.0
pyparsing==3.0.6
python-dateutil==2.8.2
pyzmq==22.3.0
six==1.16.0
